Score & Sound Library Implemented
Need to adjust font size of entire HUD.  The first menu screen wasn't
present when I started working on the project this morning, but hitting
'1' still works.  Tried using the DirectXTK built-in Audio, but it was a
bitch to get working.  I used a plugin called irrKlang instead, and is
pretty easy to use.  Currently the only sound effects are explosions
when the ship collides with asteroids.