What I have
This is what I have for normal mapping. I thought it was done, but it
turns out I've been working in a shader that isn't even called anymore.
So I had to port it over to a new shader. Furthermore, 4 of my 5 group
projects were due today, so I can redirect my attention to this project.
I'm going to talk to Chris about it Thursday.